Moved existing code into appropriate packages
